// src/components/pages/services/service-list/types.ts
export interface ServiceData {
  config_dir: string;
  config_options: string;
  data_dir: string;
  dsn: string;
  exposed_ports: string;
  id: string;
  internal_ip: string;
  links: string;
  service_root: string;
  status: string;
  version: string;
  plugin_name: string;
  initial_network?: string;
  post_create_network?: string;
  post_start_network?: string;
}

export interface ServiceListItem {
  pluginType: string;
  serviceName: string;
  serviceData: ServiceData | null;
  loading: boolean;
  error: string | null;
}